+++ pkgs/dart_tooling_mcp_server/lib/src/utils/cli_utils.dart
@@ -1,0 +1,94 @@
+// Copyright (c) 2025, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'dart:io';
+
+import 'package:dart_mcp/server.dart';
+import 'package:process/process.dart';
+
+/// Helper to run a command in multiple project roots.
+///
+/// [defaultPaths] may be specified if one or more path arguments are required
+/// for the command (e.g. `dart format <default paths>`).
+Future<CallToolResult> runCommandInRoots(
+  CallToolRequest request, {
+  required List<String> command,
+  required String commandDescription,
+  required ProcessManager processManager,
+  List<String> defaultPaths = const <String>[],
+}) async {
+  final rootConfigs =
+      (request.arguments?['roots'] as List?)?.cast<Map<String, Object?>>();
+  if (rootConfigs == null) {
+    return CallToolResult(
+      content: [TextContent(text: 'Missing required argument `roots`.')],
+      isError: true,
+    );
+  }
+
+  final outputs = <TextContent>[];
+  for (var rootConfig in rootConfigs) {
+    final rootUriString = rootConfig['root'] as String?;
+    if (rootUriString == null) {
+      // This shouldn't happen based on the schema, but handle defensively.
+      return CallToolResult(
+        content: [
+          TextContent(text: 'Invalid root configuration: missing `root` key.'),
+        ],
+        isError: true,
+      );
+    }
+
+    final rootUri = Uri.parse(rootUriString);
+    if (rootUri.scheme != 'file') {
+      return CallToolResult(
+        content: [
+          TextContent(
+            text:
+                'Only file scheme uris are allowed for roots, but got '
+                '$rootUri',
+          ),
+        ],
+        isError: true,
+      );
+    }
+    final projectRoot = Directory(rootUri.toFilePath());
+
+    final paths = (rootConfig['paths'] as List?)?.cast<String>();
+    if (paths != null) {
+      command.addAll(paths);
+    } else {
+      command.addAll(defaultPaths);
+    }
+
+    final result = await processManager.run(
+      command,
+      workingDirectory: projectRoot.path,
+      runInShell: true,
+    );
+
+    final output = (result.stdout as String).trim();
+    final errors = (result.stderr as String).trim();
+    if (result.exitCode != 0) {
+      return CallToolResult(
+        content: [
+          TextContent(
+            text:
+                '$commandDescription failed in ${projectRoot.path}:\n'
+                '$output\n\nErrors\n$errors',
+          ),
+        ],
+        isError: true,
+      );
+    }
+    if (output.isNotEmpty) {
+      outputs.add(
+        TextContent(
+          text: '$commandDescription in ${projectRoot.path}:\n$output',
+        ),
+      );
+    }
+  }
+  return CallToolResult(content: outputs);
+}
