+++ pkgs/dart_tooling_mcp_server/test/tools/dart_cli_test.dart
@@ -32,7 +32,6 @@ void main() {
   group('dart cli tools', () {
     late Tool dartFixTool;
     late Tool dartFormatTool;
-    late Tool dartPubTool;
 
     setUp(() async {
       final tools = (await testHarness.mcpServerConnection.listTools()).tools;
@@ -41,9 +40,6 @@ void main() {
       );
       dartFormatTool = tools.singleWhere(
         (t) => t.name == DartCliSupport.dartFormatTool.name,
-      );
-      dartPubTool = tools.singleWhere(
-        (t) => t.name == DartCliSupport.dartPubTool.name,
       );
     });
 
@@ -102,158 +98,6 @@ void main() {
       expect(testProcessManager.commandsRan, [
         ['dart', 'format', 'foo.dart', 'bar.dart'],
       ]);
-    });
-
-    group('dart pub', () {
-      test('add', () async {
-        final request = CallToolRequest(
-          name: dartPubTool.name,
-          arguments: {
-            'command': 'add',
-            'packageName': 'foo',
-            'roots': [
-              {'root': testRoot.uri},
-            ],
-          },
-        );
-        final result = await testHarness.callToolWithRetry(request);
-
-        // Verify the command was sent to the process maanger without error.
-        expect(result.isError, isNot(true));
-        expect(testProcessManager.commandsRan, [
-          ['dart', 'pub', 'add', 'foo'],
-        ]);
-      });
-
-      test('remove', () async {
-        final request = CallToolRequest(
-          name: dartPubTool.name,
-          arguments: {
-            'command': 'remove',
-            'packageName': 'foo',
-            'roots': [
-              {'root': testRoot.uri},
-            ],
-          },
-        );
-        final result = await testHarness.callToolWithRetry(request);
-
-        // Verify the command was sent to the process maanger without error.
-        expect(result.isError, isNot(true));
-        expect(testProcessManager.commandsRan, [
-          ['dart', 'pub', 'remove', 'foo'],
-        ]);
-      });
-
-      test('get', () async {
-        final request = CallToolRequest(
-          name: dartPubTool.name,
-          arguments: {
-            'command': 'get',
-            'roots': [
-              {'root': testRoot.uri},
-            ],
-          },
-        );
-        final result = await testHarness.callToolWithRetry(request);
-
-        // Verify the command was sent to the process maanger without error.
-        expect(result.isError, isNot(true));
-        expect(testProcessManager.commandsRan, [
-          ['dart', 'pub', 'get'],
-        ]);
-      });
-
-      test('upgrade', () async {
-        final request = CallToolRequest(
-          name: dartPubTool.name,
-          arguments: {
-            'command': 'upgrade',
-            'roots': [
-              {'root': testRoot.uri},
-            ],
-          },
-        );
-        final result = await testHarness.callToolWithRetry(request);
-
-        // Verify the command was sent to the process maanger without error.
-        expect(result.isError, isNot(true));
-        expect(testProcessManager.commandsRan, [
-          ['dart', 'pub', 'upgrade'],
-        ]);
-      });
-
-      group('returns error', () {
-        test('for missing command', () async {
-          final request = CallToolRequest(name: dartPubTool.name);
-          final result = await testHarness.callToolWithRetry(
-            request,
-            expectError: true,
-          );
-
-          expect(
-            (result.content.single as TextContent).text,
-            'Missing required argument `command`.',
-          );
-          expect(testProcessManager.commandsRan, isEmpty);
-        });
-
-        test('for unsupported command', () async {
-          final request = CallToolRequest(
-            name: dartPubTool.name,
-            arguments: {'command': 'publish'},
-          );
-          final result = await testHarness.callToolWithRetry(
-            request,
-            expectError: true,
-          );
-
-          expect(
-            (result.content.single as TextContent).text,
-            contains('Unsupported pub command `publish`.'),
-          );
-          expect(testProcessManager.commandsRan, isEmpty);
-        });
-
-        for (final command in SupportedPubCommand.values.where(
-          (c) => c.requiresPackageName,
-        )) {
-          test('for missing package name: $command', () async {
-            final request = CallToolRequest(
-              name: dartPubTool.name,
-              arguments: {'command': command.name},
-            );
-            final result = await testHarness.callToolWithRetry(
-              request,
-              expectError: true,
-            );
-
-            expect(
-              (result.content.single as TextContent).text,
-              'Missing required argument `packageName` for the '
-              '`${command.name}` command.',
-            );
-            expect(testProcessManager.commandsRan, isEmpty);
-          });
-        }
-
-        test('for missing roots', () async {
-          final request = CallToolRequest(
-            name: dartPubTool.name,
-            arguments: {'command': 'get'},
-          );
-          final result = await testHarness.callToolWithRetry(
-            request,
-            expectError: true,
-          );
-
-          expect(
-            (result.content.single as TextContent).text,
-            'Missing required argument `roots`.',
-          );
-          expect(testProcessManager.commandsRan, isEmpty);
-        });
-      });
     });
   });
 }
